requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.129.0
flatbuffers==25.12.19
gast==0.7.0
google-pasta==0.2.0
grpcio==1.78.1
h11==0.16.0
h5py==3.15.1
idna==3.11
keras==3.13.2
libclang==18.1.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.1.3
opt_einsum==3.4.0
optree==0.18.0
packaging==26.0
pillow==12.1.1
protobuf==5.29.6
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-multipart==0.0.22
requests==2.32.5
rich==14.3.3
six==1.17.0
starlette==0.52.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
Werkzeug==3.1.6
wrapt==2.1.1
